Redux set up + GetList + PostList + refactoring de la structure
