+++ frontend/src/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App/App';
-import * as serviceWorker from './serviceWorker';
+import App from './react/Components/App/App';
 import { Provider } from 'react-redux'
 
-import  store from './store'
+import  store from './redux/store'
 
 ReactDOM.render(
     <Provider store={store}>
@@ -13,7 +11,4 @@ ReactDOM.render(
     </Provider>
     , document.getElementById('root'));
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
+
