+++ frontend/src/redux/list/constants.js
@@ -1,2 +1,4 @@
-export const FETCH_LIST = 'FETCH_LIST'
-export const POST_LIST = 'POST_LIST'
+export const ACTIONS = {
+    SET_LISTS : 'SET_LISTS',
+    POST_LIST : 'POST_LIST',
+}
