+++ frontend/src/redux/reducers.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux'
-import listReducer from './listReducer'
+import listReducer from './list/reducer'
 
 export default combineReducers({
     lists : listReducer,
